test: migrate test_materialized_partition_columns_excluded_from_stats to kernel builder
Moves the test from kernel/tests/write/append.rs to its proper topic
module kernel/tests/write/partitioned.rs, and switches its setup from
the test_utils::create_table JSON helper (which hard-codes
writerFeatures = ["materializePartitionColumns"]) to the kernel
create_table builder using
delta.feature.materializePartitionColumns=supported.

The migration removes the hand-rolled protocol from the test setup so
the same path that connectors will use to enable the feature is
exercised end-to-end. Test assertions are unchanged.

Refs #2478.
